requirements.txt
typer[all]==0.12.3
rich==13.7.1
openai==1.14.3
python-dotenv==1.0.1
pyyaml==6.0.1
